Conform better to RFC3986
The RFC states that names must begin or end with an alphanumeric character, however the regex here also allows other gen_delims and sub_delims, except -, as that is not allowed in DNS and fails when calling idn_to_ascii.

Modifed the test containing com-, added some others with dashes that are expected fails, as well as adding a few other with gen_delims and sub_delims.